requirements.txt
Django==1.11.6
#django-tagging==0.3.1
django-tagging==0.4.6
django-sendfile==0.3.11
lxml==3.4.4
django-taggit==0.22.1
